sock: Add a zero copy receive interface
This is not implemented by any sock module just yet.

The basic concept is that the user provides buffers to a socket group.
These get filled in with the next portions of the stream on each socket.
The user then calls spdk_sock_recv_next() to get the pointer to the
buffer holding the next part of the stream. When the user is done, they
can put the buffer back to the group to be used again.

The provided buffers are held in a pool on the socket group.
Implementations can request a buffer from the pool or set up
a notification any time a buffer is returned.

Change-Id: I236fca43c6e3e11aafeb58caba0e4798654124be
Signed-off-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/16105
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <james.r.harris@intel.com>