test/nvmf: Verify perf's exit status
Since the perf application is run in the background the script would
not pick up its potential crash - the main condition itself was simply
checking if app is still running. Since there were occasions where
the failure was missed up until the cleanup stage, try to catch it
early on by verifying perf's exit status - hook it into NOT() as
the app is expected to fail by the test but we need to filter out
crash scenarios.

Signed-off-by: Michal Berger <michal.berger@intel.com>
Change-Id: I9a9304616c0a4dba42467f65c1a2c770352f08ea
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/16345
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>